+++ BitmovinConvivaAnalytics/Classes/ConvivaAnalytics.swift
@@ -333,6 +333,14 @@ public final class ConvivaAnalytics: NSObject {
         contentMetadataBuilder.streamUrl = player.source?.sourceConfig.url.absoluteString
     }
 
+    private func customEvent(event: PlayerViewEvent, args: [String: String] = [:]) {
+        if !isSessionActive {
+            return
+        }
+
+        sendCustomPlaybackEvent(name: event.name, attributes: args)
+    }
+
     private func customEvent(event: PlayerEvent, args: [String: String] = [:]) {
         if !isSessionActive {
             return
@@ -341,7 +349,7 @@ public final class ConvivaAnalytics: NSObject {
         sendCustomPlaybackEvent(name: event.name, attributes: args)
     }
 
-    private func onPlaybackStateChanged(playerState: PlayerState) {
+    private func onPlaybackStateChanged(playerState: ConvivaSDK.PlayerState) {
         // do not report any playback state changes while player isStalled except buffering
         if isStalled && playerState != .CONVIVA_BUFFERING {
             return
@@ -357,6 +365,15 @@ public final class ConvivaAnalytics: NSObject {
         videoAnalytics.reportPlaybackMetric(CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE, value: playerState.rawValue)
         logger.debugLog(message: "Player state changed: \(playerState.rawValue)")
     }
+
+    private func reportPlayHeadTime() {
+        guard isSessionActive else { return }
+
+        videoAnalytics.reportPlaybackMetric(
+            CIS_SSDK_PLAYBACK_METRIC_PLAY_HEAD_TIME,
+            value: Int64(player.currentTime(.relativeTime) * 1000)
+        )
+    }
 }
 
 // MARK: - PlayerListener
@@ -370,6 +387,7 @@ extension ConvivaAnalytics: BitmovinPlayerListenerDelegate {
     }
 
     func onTimeChanged() {
+        reportPlayHeadTime()
         updateSession()
     }
 
@@ -484,7 +502,10 @@ extension ConvivaAnalytics: BitmovinPlayerListenerDelegate {
     #if !os(tvOS)
     // MARK: - Ad events
     func onAdStarted(_ event: AdStartedEvent) {
-        let adPosition: AdPosition = AdEventUtil.parseAdPosition(event: event, contentDuration: player.duration)
+        let adPosition: ConvivaSDK.AdPosition = AdEventUtil.parseAdPosition(
+            event: event,
+            contentDuration: player.duration
+        )
         var adAttributes = [String: Any]()
         adAttributes["c3.ad.position"] = adPosition.rawValue
         videoAnalytics.reportAdBreakStarted(AdPlayer.ADPLAYER_CONTENT,