+++ Example/Tests/ContentMetadataTests.swift
@@ -62,11 +62,11 @@ class ContentMetadataSpec: QuickSpec {
                     let spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "reportPlaybackRequested")
 
                     let playerConfig = PlayerConfig()
-                    _ = TestDouble(aClass: playerDouble, name: "config", return: playerConfig)
+                    _ = TestDouble(aClass: playerDouble!, name: "config", return: playerConfig)
 
                     let sourceConfig = SourceConfig(url: URL(string: "www.google.com.m3u8")!, type: .hls)
                     sourceConfig.title = "Art of Unit Test"
-                    let source = SourceFactory.create(from: sourceConfig)
+                    let source = SourceFactory.createSource(from: sourceConfig)
                     playerDouble.load(source: source)
                     playerDouble.fakePlayEvent() // to initialize session
                     expect(spy).to(
@@ -97,10 +97,10 @@ class ContentMetadataSpec: QuickSpec {
                     let playerConfig = PlayerConfig()
                     let sourceConfig = SourceConfig(url: URL(string: "www.google.com.m3u8")!, type: .hls)
                     sourceConfig.title = "Art of Unit Test"
-                    let source = SourceFactory.create(from: sourceConfig)
+                    let source = SourceFactory.createSource(from: sourceConfig)
 
-                    _ = TestDouble(aClass: playerDouble, name: "config", return: playerConfig)
-                    _ = TestDouble(aClass: playerDouble, name: "source", return: source)
+                    _ = TestDouble(aClass: playerDouble!, name: "config", return: playerConfig)
+                    _ = TestDouble(aClass: playerDouble!, name: "source", return: source)
 
                     playerDouble.fakePlayEvent() // to initialize session
 
@@ -141,10 +141,10 @@ class ContentMetadataSpec: QuickSpec {
                     let playerConfig = PlayerConfig()
                     let sourceConfig = SourceConfig(url: URL(string: "www.google.com.m3u8")!, type: .hls)
                     sourceConfig.title = "Art of Unit Test"
-                    let source = SourceFactory.create(from: sourceConfig)
+                    let source = SourceFactory.createSource(from: sourceConfig)
 
-                    _ = TestDouble(aClass: playerDouble, name: "config", return: playerConfig)
-                    _ = TestDouble(aClass: playerDouble, name: "source", return: source)
+                    _ = TestDouble(aClass: playerDouble!, name: "config", return: playerConfig)
+                    _ = TestDouble(aClass: playerDouble!, name: "source", return: source)
 
                     playerDouble.fakePlayEvent() // to initialize session
 
@@ -152,7 +152,7 @@ class ContentMetadataSpec: QuickSpec {
                         haveBeenCalled(withArgs: ["streamUrl": "www.google.com.m3u8"])
                     )
                 }
- 
+
                 it("update bitrate") {
                      let videoQuality = VideoQuality(identifier: "Test",
                                                      label: "test",
@@ -161,7 +161,7 @@ class ContentMetadataSpec: QuickSpec {
                                                      width: 1900,
                                                      height: 800)
 
-                    _ = TestDouble(aClass: playerDouble, name: "videoQuality", return: videoQuality)
+                    _ = TestDouble(aClass: playerDouble!, name: "videoQuality", return: videoQuality)
 
                     playerDouble.fakePlayEvent() // to initialize session
                     let spy = Spy(aClass: CISVideoAnalyticsTestDouble.self, functionName: "reportPlaybackMetric")
@@ -287,10 +287,10 @@ class ContentMetadataSpec: QuickSpec {
 
                         let sourceConfig = SourceConfig(url: URL(string: "http://a.url")!, type: .hls)
                         sourceConfig.title = "MyTitle"
-                        let source = SourceFactory.create(from: sourceConfig)
+                        let source = SourceFactory.createSource(from: sourceConfig)
                         playerDouble.load(source: source)
 
-                        _ = TestDouble(aClass: playerDouble, name: "config", return: playerConfig)
+                        _ = TestDouble(aClass: playerDouble!, name: "config", return: playerConfig)
 
                         playerDouble.fakePlayEvent()
                         playerDouble.fakePlayingEvent()