+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+## [3.1.0] - 2024-05-27
+
+### Added
+- Hook into `onTimeChanged` callback for reporting Play head time (PHT) to Conviva playback metric. 
+
+### Changed
+- Raised minimum deployment targets to iOS/tvOS 14.0
+- Updated minimum Bitmovin Player version to 3.63.0
+
+### Removed
+- iOS/tvOS 12 and 13 support
+
 ## [3.0.1]
 
 ### Fixed